KendoAI/hello_world/requirements.txt
boto3
requests
line-bot-sdk